% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{bkgrnd_corr}
\alias{bkgrnd_corr}
\title{Subset common peptide sequences}
\usage{
bkgrnd_corr(substrates_dt)
}
\arguments{
\item{substrates_dt}{An enzymatic preference data.table imported using `import_substrates()`}
}
\value{
A data.table containing sequences shared in the maximum number of 
experimental replicates and minimum number of control replicates 
(using the `replicate` and `control` columns).
}
\description{
Subset common peptide sequences
}
\examples{
uniq_subs <- bkgrnd_corr(substrates)
}
