% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_signature.R
\name{get_candidateaa}
\alias{get_candidateaa}
\title{Generate candidate position-specific amino acid residues}
\usage{
get_candidateaa(fisher_tables, screener_dt, target_kinase)
}
\arguments{
\item{fisher_tables}{A list containing two data tables: 1) Exact Test P-values
and 2) Fisher Exact Odds.}

\item{screener_dt}{A four-element list generated using the
`multi_screener()` function.}

\item{target_kinase}{Abbreviation (typically 3 letter) for the enzyme the 
substrate is to be optimized for.}
}
\value{
A data.table containing candidate amino acid residues and associated 
p-value and odds from Fisher Exact Test.
}
\description{
Generate candidate position-specific amino acid residues
}
