% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_signature.R
\name{get_normscores}
\alias{get_normscores}
\title{Scaling raw peptide scores from candidate list}
\usage{
get_normscores(candidates, screener_dt, n_hits = NULL)
}
\arguments{
\item{candidates}{List of candidate sequence information created using
`generate_substrates()`}

\item{screener_dt}{A four-element list generated using the
`multi_screener()` function.}

\item{n_hits}{Number of candidate sequence scores to be re-scaled.}
}
\value{
A two-element list containing 1) Scaled scores for each substrate 
and 2) Scaled cut-off scores for activity with a particular enzyme.
}
\description{
Scaling raw peptide scores from candidate list
}
