% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screener.R
\name{make_screener}
\alias{make_screener}
\title{Create new input screener file}
\usage{
make_screener()
}
\value{
A data.table containing the substrate preferences for all 
enzyme preference files in a selected folder.
}
\description{
Create new input screener file
}
