% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{multi_sample}
\alias{multi_sample}
\title{Import enzyme substrate preference files and parse sample type}
\usage{
multi_sample(
  path,
  ptm_type = c("Oxidation (M)-M", "Carbamidomethylation-C", "Deamidation (NQ)-N",
    "Deamidation (NQ)-Q", "Phosphorylation (STY)-Y", "Phosphorylation (STY)-T",
    "Phosphorylation (STY)-S", "Acetylation (Protein N-term)-E",
    "Acetylation (Protein N-term)-D", "Acetylation (Protein N-term)-M",
    "Acetylation (Protein N-term)-S", "Acetylation (Protein N-term)-A",
    "Acetylation (Protein N-term)-T", "Acetylation (Protein N-term)-G",
    "Acetylation (Protein N-term)-C", "Acetylation (Protein N-term)-V",
    "Pyro-glu from Q-Q"),
  legacy = FALSE
)
}
\arguments{
\item{path}{The directory containing the analyzed phosphoproteomics data}

\item{ptm_type}{The post-translational modification to be analyzed and its 
targeted the amino acid. Currently, only one ptm type can be analyzed at a 
time. A list of ptm options can be accessed by typing `ptm_key`}

\item{legacy}{A logical parameter where TRUE is for files generated using the
Galaxy-P KinaMine workflow. Is set to FALSE by default.}
}
\value{

}
\description{
Import enzyme substrate preference files and parse sample type
}
