% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screener.R
\name{multi_screener}
\alias{multi_screener}
\title{Generate enzyme screener file}
\usage{
multi_screener(
  screener_input,
  uniprot_input,
  path,
  method = c("prod", "log2_sum", "w_prod"),
  pval_corr = FALSE,
  type = "aa",
  norm_method = c("none", "bkgrnd"),
  property = NULL,
  constrain = NULL
)
}
\arguments{
\item{screener_input}{An enzyme-specific preferred substrates file.}

\item{uniprot_input}{The `uniprot_id`-matched in silico control peptide library.}

\item{path}{The file path where output data should be saved.}

\item{method}{Peptide scoring algorithm. Accepts product of odds ratios 
(`prod`), weighted product of odds ratio (`w_prod`), and the sum of 
log scores (`log2_sum`).}

\item{pval_corr}{A logical parameter where TRUE will set all fisher odds 
with non-significant p-values (> 0.05) to 1. Set to FALSE by default.}

\item{type}{Character of either `aa` or `aa_property` to indicate whether 
amino acid residues or amino acid properties should be used.}

\item{norm_method}{Character of either `none` or `bkgrnd` to indicate whether 
raw scores or background-corrected ones should be used. Background-correction
subtracts the mean and divides by the standard deviation of the negative 
control substrate scores.}

\item{property}{Character indicating which amino acid property should be 
analyzed. Only used if type = `aa`. See available property options 
by typing `aa_classificiation`.}

\item{constrain}{The minimum specificity required when calculating 
the threshold score for activity. By default set to 0.9.}
}
\value{
A three-element list containing: 1) A lot of score distributions and
the selected threshold, 2) Fisher tables for each kinase in the 
`screener_input` file, and 3) Summary stats on score distribution for 
each enzyme.
}
\description{
The function multi_screener takes as inputs 1) 
the kinase-specific preferred substrates file and 2) the protein-matched 
in silico control peptide library. For each enzyme, an optimal threshold will
be calculated to separate the peptide scores of sample vs control 
sequences. This is calculated using the package `cutpointr.` A bootstrap model
 is used to select an optimal cutpoint to maximize sensitivity.
}
\examples{
screener <- multi_screener(screener_raw, screener_uniprot, 
                           path = output_dir,
                           method = "prod",
                           pval_corr = FALSE,
                           type = "aa",
                           norm_method = "none",
                           constrain = 0.90)
}
