% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_signature.R
\name{norm_scores}
\alias{norm_scores}
\title{Scaling raw peptide scores from a table}
\usage{
norm_scores(scoretable, screener_dt)
}
\arguments{
\item{scoretable}{A data.table with the same structure as the 
score table generated using `score_candidates()`.}

\item{screener_dt}{A four-element list generated using the
`multi_screener()` function.}
}
\value{
A two-element list containing 1) Scaled scores for each substrate 
and 2) Scaled cut-off scores for activity with a particular enzyme.
}
\description{
Scaling raw peptide scores from a table
}
