% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{peptide_intersect}
\alias{peptide_intersect}
\title{Find peptide commonality across replicates}
\usage{
peptide_intersect(substrates_dt)
}
\arguments{
\item{substrates_dt}{An enzymatic preference data.table.}
}
\value{
A ggplot2 object.
}
\description{
Find peptide commonality across replicates
}
