% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substrate_scoring.R
\name{peptide_screener}
\alias{peptide_screener}
\title{Score an individual existing peptide sequence}
\usage{
peptide_screener(screener_dt, candidates_dt, target_kinase, family = FALSE)
}
\arguments{
\item{screener_dt}{A four-element list generated using the
`multi_screener()` function.}

\item{candidates_dt}{A data.table of candidate substrate sequences}

\item{target_kinase}{Abbreviation (typically 3 letter) for the enzyme the 
substrate is to be optimized for.}

\item{family}{A logical indicating whether to screen specifically against 
a kinase family (TRUE) or not (FALSE). The default value is FALSE.}

\item{kinase}{Abbreviation (typically 3 letter) for the enzyme the 
substrate is to be optimized for.}
}
\value{
A data.table containing substrate scoring information for each enzyme 
in the provided `screener_dt` file.
}
\description{
Score an individual existing peptide sequence
}
