% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{pssm_volcano}
\alias{pssm_volcano}
\title{Create a PSSM volcano plot}
\usage{
pssm_volcano(pssm, odds_cutoff = NULL, pval_cutoff = NULL)
}
\arguments{
\item{pssm}{A list containing two data tables: 1) Exact Test P-values
and 2) Fisher Exact Odds.}

\item{odds_cutoff}{The log 2 odds ratio threshold for being considered 
significant. If not specified it is set to 0.5 automatically.}

\item{pval_cutoff}{The threshold p-value to be considered significant. 
If not specified it is set to 0.05 automatically.}
}
\value{
A ggplot2 object
}
\description{
Create a PSSM volcano plot
}
