% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_signature.R
\name{score_candidates}
\alias{score_candidates}
\title{Score candidate artificial peptide substrate sequences}
\usage{
score_candidates(fisher_candidates, screener_dt, kinase = "ALL")
}
\arguments{
\item{fisher_candidates}{A list containing two data tables: 1) Exact Test P-values
and 2) Fisher Exact Odds.}

\item{screener_dt}{A four-element list generated using the
`multi_screener()` function.}

\item{kinase}{A vector of kinases to optimize against. 
Setting screening_kinase = "ALL" will screen against every enzyme in the 
existing screener file}
}
\value{

}
\description{
Score candidate artificial peptide substrate sequences
}
