% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pssm.R
\name{substrate_fisher_test}
\alias{substrate_fisher_test}
\title{Construct PSSM and perform Fisher Exact Test}
\usage{
substrate_fisher_test(
  substrates_dt,
  uniprot_dt,
  type = c("aa", "aa_property"),
  property = c("property_chemical", "property_hydropathy", "property_volume",
    "property_polar", "property_charge")
)
}
\arguments{
\item{substrates_dt}{An enzymatic preference data.table.}

\item{uniprot_dt}{A data.table containing in silico negative peptide 
sequences for each `uniprot_id` found in the `substrates_dt` file.}

\item{type}{Character of either `aa` or `aa_property` to indicate whether 
amino acid residues or amino acid properties should be used.}

\item{property}{Character indicating which amino acid property should be 
analyzed. Only used if type = `aa`. See available property options 
by typing `aa_classificiation`.}
}
\value{
A list containing two data tables: 1) Exact Test P-values and 
2) Fisher Exact Odds
}
\description{
Construct PSSM and perform Fisher Exact Test
}
\examples{
pssm <- substrate_fisher_test(substrates_dt = uniq_subs, uniprot_dt = uniprot,
type = "aa")
}
