% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{substrates_property}
\alias{substrates_property}
\title{Plot amino acid properties across the flank sequence}
\usage{
substrates_property(substrates_dt, uniprot_dt, path)
}
\arguments{
\item{substrates_dt}{An enzymatic preference data.table.}

\item{uniprot_dt}{A data.table containing in silico negative peptide 
sequences for each `uniprot_id` found in the `substrates_dt` file.}

\item{path}{The file path where output data should be saved.}
}
\value{
A ggplot2 object.
}
\description{
Plot amino acid properties across the flank sequence
}
