% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{substrates_qc}
\alias{substrates_qc}
\title{Generate QC report}
\usage{
substrates_qc(substrates_dt, path)
}
\arguments{
\item{substrates_dt}{An enzymatic preference data.table.}

\item{path}{The file path where output data should be saved.}
}
\value{
A ggplot2 object.
}
\description{
Generate QC report
}
